requirements.txt
torch>=2.1,<2.3
torchvision>=0.16,<0.18
ultralytics>=8.0,<9.0
opencv-python>=4.8,<4.9
pillow>=10.0,<11.0
numpy>=1.23,<2.0
pandas>=2.0,<3.0
easyocr>=1.6,<1.8
pytesseract>=0.3,<0.4
face_recognition>=1.3,<1.4
reportlab>=4.0,<4.2
ffmpeg-python>=0.2,<0.3
tqdm>=4.66,<5.0
exifread>=3.0,<4.0
pycryptodome>=3.19,<4.0
streamlit>=1.29,<1.36
motmetrics>=1.4,<1.5
imageio>=2.31,<3.0
